Correct warnings found after disabling all warning when CUDA is enabled.

+++ vtkm/worklet/spatialstructure/KdTree3DConstruction.h
@@ -58,7 +58,7 @@ public:
     template <typename T>
     VTKM_EXEC void operator()(const T& rank, const T& pointCountInSeg, T& flag) const
     {
-      if (rank >= pointCountInSeg / 2.0f)
+      if (static_cast<float>(rank) >= static_cast<float>(pointCountInSeg) / 2.0f)
         flag = 1; //right subtree
       else
         flag = 0; //left subtree
